[refactor] Removed the overkill in using BigInteger in items that will not have a lot of resources

+++ Src/Application/Interfaces/IServiceService.cs
@@ -7,7 +7,7 @@ using Application.DTOs;
 public interface IServiceService
 {
     Task<IEnumerable<ServiceResponseDto>> GetAllAsync();
-    Task<ServiceResponseDto> GetByIdAsync(BigInteger id);
+    Task<ServiceResponseDto> GetByIdAsync(int id);
     Task<ServiceResponseDto> AddAsync(ServiceRequestDto serviceRequestDto);
     Task<ServiceResponseDto> UpdateAsync(int id, ServiceRequestDto serviceRequestDto);
     Task<bool> RemoveAsync(int id);
