+++ Src/Core/Entities/Client.cs
@@ -5,7 +5,7 @@ namespace Core.Entities;
 
 public record Client
 {
-    public BigInteger Id { get; set; }
+    public int Id { get; set; }
     
     public List<Appointment> Appointments { get; set; }
     public List<Establishment> Establishments { get; set; }
